refactor to move container logic into container layer
